Use dictionary of request packet when decoding the response packet (#11)
By using the dictionary of the request packet to decode the response packet we are able to decode vendor specific attributes, should they be present in the response. The normal flow where vendor specific attributes are not present should continue to operate as usual with the default dictionary as set in the request packet by default.