Fix: Bug in preparing User-Password Attribute
This cause user with long password cannot log in

Reason:
According to section-5.2 in RFC 2865 [1]
When the password is longer than 16 characters

 c(i) = pi xor (MD5(S + c(i-1)))

So when calculating the second 16 octets of the String
field of the User-Password Attribute: c(2). It should be:

 c(2) = p2 xor (MD5(S + c(1)))

While In the current implement AccessRequest.encodePapPassword(),
the byte array 'lastBlock' keeps
 "the value of the first 16 octets segment of the password": p1.
Actually, it should keep
 "the value of the first 16 octets of the String field of the User-Password Attribute": c(1).

Fixed.

Also, Delete the loop of assigning zero value to byte array
'encryptedPass' as this is default behavior in Java.

[1] https://tools.ietf.org/html/rfc2865#section-5.2